pyproject.toml
[project]
name = "bq-caapi-wrapper-w-ge"
version = "0.1.0"
description = "ADK Agents for Conversational Analytics with Gemini Enterprise"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "google-adk>=1.23.0",
    "google-cloud-geminidataanalytics>=0.9.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]
